requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==4.0
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
whitenoise==6.0.0
